requirements.txt
kafka-python==2.0.5
opencv-python==4.10.0.84
fastapi==0.115.12
uvicorn==0.34.0
ultralytics==8.3.102
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
roboflow==1.1.48
imageio==2.37.0
imageio[ffmpeg]
av==14.3.0
minio==7.2.15 
pydantic==2.11.3
python-multipart==0.0.20
cassandra-driver==3.29.2
pydantic-settings==2.8.1
boto3==1.37.23
ffmpeg-python==0.2.0
uvicorn[standard]
onnx==1.17.0
onnxruntime==1.20.0
onnxruntime-gpu==1.20.0
tensorrt==10.8.0.43
bytetracker==0.3.2
tritonclient==2.54.0
tritonclient[all]
# pytesseract
# fast_plate_ocr
# paddleocr
# paddlepaddle